update: dependencies
