+++ composer.json
@@ -25,19 +25,19 @@
         "hootlex/laravel-moderation": "^1.1",
         "intervention/image": "^2.7.2",
         "joypixels/assets": "^6.6",
-        "laravel/framework": "^10.4.1",
+        "laravel/framework": "^10.7.1",
         "laravel/tinker": "^2.8.1",
         "laravel/ui": "^4.2.1",
-        "league/flysystem-sftp-v3": "^3.10.3",
+        "league/flysystem-sftp-v3": "^3.14",
         "livewire/livewire": "^2.12.3",
         "marcreichel/igdb-laravel": "3.7.0",
         "paragonie/constant_time_encoding": "^2.6.3",
         "predis/predis": "^2.1.2",
-        "spatie/laravel-backup": "^8.1.7",
+        "spatie/laravel-backup": "^8.1.8",
         "spatie/laravel-cookie-consent": "^3.2.4",
         "spatie/laravel-image-optimizer": "^1.7.1",
         "spatie/ssl-certificate": "^1.22.1",
-        "symfony/dom-crawler": "^6.2.7",
+        "symfony/dom-crawler": "^6.2.8",
         "theodorejb/polycast": "dev-master",
         "voku/anti-xss": "^4.1.41",
         "vstelmakh/url-highlight": "^3.0.2"
@@ -45,12 +45,12 @@
     "require-dev": {
         "brianium/paratest": "^6.9.1",
         "fakerphp/faker": "^1.21",
-        "laravel/pint": "^1.6",
-        "laravel/sail": "^1.21.2",
+        "laravel/pint": "^1.8",
+        "laravel/sail": "^1.21.4",
         "mockery/mockery": "^1.5.1",
         "nunomaduro/collision": "^6.4",
-        "phpunit/phpunit": "^9.6.5",
-        "spatie/laravel-ignition": "^2.0"
+        "phpunit/phpunit": "^9.6.6",
+        "spatie/laravel-ignition": "^2.1"
     },
     "config": {
         "preferred-install": "dist",
