+++ package.json
@@ -16,7 +16,6 @@
     "laravel-echo": "^1.11.4",
     "laravel-mix": "^6.0.43",
     "laravel-mix-purgecss": "^6.0.0",
-    "lodash": "^4.17.21",
     "postcss": "^8.4.12",
     "resolve-url-loader": "^5.0.0",
     "sass": "^1.49.9",
@@ -32,7 +31,6 @@
     "alpinejs": "^3.9.1",
     "dayjs": "^1.11.0",
     "laravel-echo-server": "^1.6.3",
-    "node-vibrant": "^3.1.6",
     "pusher-js": "^7.0.6",
     "virtual-select-plugin": "^1.0.26"
   }
