+++ resources/js/unit3d/chat.js
@@ -1,4 +1,19 @@
 /**
+ * Next, we will create a fresh Vue application instance and attach it to
+ * the page. Then, you may begin adding components to this application
+ * or customize the JavaScript scaffolding to fit your unique needs.
+ */
+import Vue from 'vue';
+import chatbox from '../components/chat/Chatbox';
+
+if (document.getElementById('vue')) {
+    new Vue({
+        el: '#vue',
+        components: { chatbox: chatbox },
+    });
+}
+
+/**
  * Echo exposes an expressive API for subscribing to channels and listening
  * for events that are broadcast by Laravel. Echo and event broadcasting
  * allows your team to easily build robust real-time web applications.
